requirements.txt
bcrypt==2.0.0
cffi==1.4.2
cryptography==1.2.1
Django==1.7.4
django-cors-headers==1.0.0
django-filter==0.9.1
djangorestframework==3.3.2
enum34==1.1.2
idna==2.0
ipaddress==1.0.16
Markdown==2.5.2
Pillow==3.0.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
python-memcached==1.57
requests==2.9.1
six==1.10.0
solrpy==0.9.7
wheel==0.24.0
-e git+https://github.com/scieloorg/xylose@1.4.4#egg=xylose
